[Bronze V] Title: 이상한 기호, Time: 168 ms, Memory: 17648 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -4,8 +4,8 @@ public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         String[] input = sc.nextLine().split(" ");
-        int A = Integer.parseInt(input[0]);
-        int B = Integer.parseInt(input[1]);
+        Long A = Long.parseLong(input[0]);
+        Long B = Long.parseLong(input[1]);
         
         System.out.println((A+B)*(A-B));
         sc.close();
